+++ web/src/views/library/Library.vue
@@ -76,7 +76,7 @@ function handleTabChange(tabId: string | number) {
         @update:model-value="handleTabChange"
         class="w-full h-full flex flex-col"
       >
-        <div class="-mx-3 px-3 flex items-end border-b relative" style="width: calc(100% + 1.5rem)">
+        <div class="-mx-3 px-3 flex items-end border-b" style="width: calc(100% + 1.5rem)">
           <TabsList variant="linear" class="border-b-0">
             <TabsTrigger
               v-for="tab in tabs"
@@ -88,7 +88,6 @@ function handleTabChange(tabId: string | number) {
               {{ tab.label }}
             </TabsTrigger>
           </TabsList>
-          <div id="library-tab-actions" class="absolute right-3 bottom-1 flex items-center gap-0.5" />
         </div>
 
         <div class="flex-1 pt-3">